created a new library (jsonParseV2) with all the new JSON parse functions
