+++ src/lib/jsonParseV2/parseContextAttribute.cpp
@@ -27,11 +27,11 @@
 #include "logMsg/logMsg.h"
 
 #include "ngsi/ContextAttribute.h"
-#include "parse/jsonParseTypeNames.h"
 #include "parse/CompoundValueNode.h"
-#include "parse/parseContextAttribute.h"
-#include "parse/parseMetadata.h"
-#include "parse/parseContextAttributeCompoundValue.h"
+#include "jsonParseV2/jsonParseTypeNames.h"
+#include "jsonParseV2/parseContextAttribute.h"
+#include "jsonParseV2/parseMetadata.h"
+#include "jsonParseV2/parseContextAttributeCompoundValue.h"
 
 using namespace rapidjson;
