+++ src/lib/jsonParseV2/parseContextAttributeCompoundValue.h
@@ -1,3 +1,6 @@
+#ifndef SRC_LIB_JSONPARSEV2_PARSECONTEXTATTRIBUTECOMPOUNDVALUE_H_
+#define SRC_LIB_JSONPARSEV2_PARSECONTEXTATTRIBUTECOMPOUNDVALUE_H_
+
 /*
 *
 * Copyright 2015 Telefonica Investigacion y Desarrollo, S.A.U
@@ -23,35 +26,24 @@
 * Author: Ken Zangelin
 */
 #include <string>
-#include <vector>
 
-#include "rest/ConnectionInfo.h"
-#include "ngsi/ParseData.h"
-#include "ngsi/Request.h"
-#include "parse/jsonRequestTreat.h"
-#include "parse/parseEntity.h"
+#include "rapidjson/document.h"
+
+#include "ngsi/ContextAttribute.h"
+
+using namespace rapidjson;
 
 
 
 /* ****************************************************************************
 *
-* jsonRequestTreat - 
+* parseContextAttribute - 
 */
-std::string jsonRequestTreat(ConnectionInfo* ciP, ParseData* parseDataP, RequestType requestType)
-{
-  std::string answer;
+extern std::string parseContextAttributeCompoundValue
+(
+  const Value::ConstMemberIterator&  node,
+  ContextAttribute*                  caP,
+  orion::CompoundValueNode*          parent
+);
 
-  LM_M(("KZ: requestType: %d", requestType));
-  switch (requestType)
-  {
-  case EntitiesRequest:
-    answer = parseEntity(ciP, &parseDataP->ent.res);
-    break;
-
-  default:
-    answer = "Request Treat function not implemented";
-    break;
-  }
-  
-  return answer;
-}
+#endif  // SRC_LIB_JSONPARSEV2_PARSECONTEXTATTRIBUTECOMPOUNDVALUE_H_
